:shirt: #14 メッセージの表示部分を別Componentに分離
